Support historyMode in SetURLParameterAction (#188)
* Support historyMode in SetURLParameterAction

To allow more flexibility when updating path/query parameters, we want to allow setting whether the action should push or replace the history state. This is also useful when you want to update multiple parameters - although we should have a separate/similar action for that scenario.

* Skip lazy evaluation of state

Since we use an if/else (and not if/else if), the state will always be evaluated atm

* Bump version